+++ src/layers/api_utils/api_step_chain/__init__.py
@@ -47,3 +47,40 @@ def execute_step_chain(
     )
     response_chain.run(init=(result, version, None))
     return response_chain.result
+
+
+def execute_step_chain_with_location(
+    event: dict, cache: dict, versioned_steps: dict[str, ModuleType]
+):
+    """After FHIR removed from everywhere can also remove this"""
+    event["headers"] = lower_case_keys(event.get("headers", {}))
+
+    version_chain = StepChain(
+        step_chain=versioning_steps, step_decorators=STEP_DECORATORS
+    )
+    version_chain.run(
+        init={
+            VERSIONING_STEP_ARGS.EVENT: event,
+            VERSIONING_STEP_ARGS.VERSIONED_STEPS: versioned_steps,
+        }
+    )
+
+    version = None
+    location = None
+    if isinstance(version_chain.result, Exception):
+        result = version_chain.result
+    else:
+        version = version_chain.data[get_largest_possible_version]
+        steps = version_chain.data[get_steps_for_requested_version]
+        api_chain = StepChain(step_chain=steps, step_decorators=STEP_DECORATORS)
+        api_chain.run(cache=cache, init=event)
+        if isinstance(api_chain.result, Exception):
+            result = api_chain.result
+        else:
+            result, location = api_chain.result
+
+    response_chain = StepChain(
+        step_chain=response_steps, step_decorators=STEP_DECORATORS
+    )
+    response_chain.run(init=(result, version, location))
+    return response_chain.result