+++ src/layers/domain/core/cpm_product/v1.py
@@ -1,18 +1,16 @@
 from datetime import datetime
-from typing import Optional
 from uuid import UUID
 
 from attr import dataclass
 from domain.core.aggregate_root import AggregateRoot
+from domain.core.cpm_system_id.v1 import ProductId
+from domain.core.device.v2 import UPDATED_ON, event
+from domain.core.error import DuplicateError
 from domain.core.event import Event
+from domain.core.product_key.v1 import ProductKey
 from domain.core.timestamp import now
-from domain.core.validation import PRODUCT_NAME_REGEX
-from pydantic import BaseModel, Extra, Field
-
-
-class CpmProductIncomingParams(BaseModel, extra=Extra.forbid):
-    product_team_id: UUID = Field(...)
-    product_name: str = Field(regex=PRODUCT_NAME_REGEX, min_length=1)
+from domain.core.validation import CPM_PRODUCT_NAME_REGEX
+from pydantic import Field
 
 
 @dataclass(kw_only=True, slots=True)
@@ -22,8 +20,21 @@ class CpmProductCreatedEvent(Event):
     name: str
     ods_code: str
     created_on: str
-    updated_on: Optional[str] = None
-    deleted_on: Optional[str] = None
+    updated_on: str = None
+    deleted_on: str = None
+
+
+@dataclass(kw_only=True, slots=True)
+class CpmProductKeyAddedEvent(Event):
+    new_key: ProductKey
+    id: str
+    product_team_id: UUID
+    name: str
+    ods_code: str
+    created_on: str
+    updated_on: str = None
+    deleted_on: str = None
+    keys: list[ProductKey]
 
 
 class CpmProduct(AggregateRoot):
@@ -31,10 +42,23 @@ class CpmProduct(AggregateRoot):
     A product in the database.
     """
 
-    id: str  # product_id
+    id: ProductId = Field(default_factory=ProductId.create)
     product_team_id: UUID = Field(...)
-    name: str = Field(regex=PRODUCT_NAME_REGEX, min_length=1)
+    name: str = Field(regex=CPM_PRODUCT_NAME_REGEX, min_length=1)
     ods_code: str
     created_on: datetime = Field(default_factory=now, immutable=True)
-    updated_on: Optional[datetime] = Field(default=None)
-    deleted_on: Optional[datetime] = Field(default=None)
+    updated_on: datetime = Field(default=None)
+    deleted_on: datetime = Field(default=None)
+    keys: list[ProductKey] = Field(default_factory=list)
+
+    @event
+    def add_key(self, key_type: str, key_value: str) -> CpmProductKeyAddedEvent:
+        product_key = ProductKey(key_value=key_value, key_type=key_type)
+        if product_key in self.keys:
+            raise DuplicateError(
+                f"It is forbidden to supply duplicate keys: '{key_type}':'{key_value}'"
+            )
+        self.keys.append(product_key)
+        product_data = self.state()
+        product_data.pop(UPDATED_ON)  # The @event decorator will handle updated_on
+        return CpmProductKeyAddedEvent(new_key=product_key, **product_data)