+++ src/layers/domain/repository/transaction.py
@@ -1,9 +1,11 @@
 from contextlib import contextmanager
 from enum import StrEnum
+from functools import partial
 from typing import Literal, Optional
 
 from botocore.exceptions import ClientError
 from domain.core.error import NotFoundError
+from domain.repository.marshall import marshall_value
 from pydantic import BaseModel, Field
 
 from .errors import AlreadyExistsError, UnhandledTransaction
@@ -86,3 +88,41 @@ def handle_client_errors(commands: list[TransactItem]):
             code=response.Error.Code,
             unhandled_transactions=commands,
         )
+
+
+def _update_expression(updated_fields: dict) -> dict:
+    expression_attribute_names = {}
+    expression_attribute_values = {}
+    update_clauses = []
+
+    for field_name, value in updated_fields.items():
+        field_name_placeholder = f"#{field_name}"
+        field_value_placeholder = f":{field_name}"
+
+        update_clauses.append(f"{field_name_placeholder} = {field_value_placeholder}")
+        expression_attribute_names[field_name_placeholder] = field_name
+        expression_attribute_values[field_value_placeholder] = marshall_value(value)
+
+    update_expression = "SET " + ", ".join(update_clauses)
+
+    return dict(
+        UpdateExpression=update_expression,
+        ExpressionAttributeNames=expression_attribute_names,
+        ExpressionAttributeValues=expression_attribute_values,
+    )
+
+
+def update_transactions(
+    table_name: str, primary_keys: list[dict], data: dict
+) -> list[TransactItem]:
+    update_expression = _update_expression(updated_fields=data)
+    update_statement = partial(
+        TransactionStatement,
+        TableName=table_name,
+        ConditionExpression=ConditionExpression.MUST_EXIST,
+        **update_expression,
+    )
+    transact_items = [
+        TransactItem(Update=update_statement(Key=key)) for key in primary_keys
+    ]
+    return transact_items